Fix files locations
